Add support for preprocessing the VolumeSecret volumes
VolumeSecret volumes refer to secrets within their definitions,
so that the resulting YAML is equivalent yet less verbose.
This requires a preprocessing of those volumes, since their secrets
are not defined in the yaml.Manifest. This commit adds a local
transform func 'withVolumeSecretFunc' to support this preprocessing.

This might better be in yaml's transform package, but it is the
first PoC of VolumeSecrets to know it yet.